[CPML] Changed n_points type from int to size_t
Updated to use size_t for expressing point number: it is the only proper
way. Changed the implementation to return 0 instead of -1 on errors and
updated the library accordingly.

The n_point arg in cpml_primitive_get_point() is still an int because
negative offsets are accepted.

+++ cpml/cpml-primitive-private.h
@@ -56,7 +56,7 @@ typedef struct __CpmlPrimitiveClass _CpmlPrimitiveClass;
 
 struct __CpmlPrimitiveClass {
     const char   *name;
-    int           n_points;
+    size_t        n_points;
 
     double       (*get_length)          (const CpmlPrimitive    *primitive);
     void         (*put_extents)         (const CpmlPrimitive    *primitive,
